rename templates package to markup
