+++ go.mod
@@ -3,6 +3,7 @@ module github.com/facundoolano/jorge
 go 1.22.0
 
 require (
+	github.com/alecthomas/kong v0.8.1
 	github.com/elliotchance/orderedmap/v2 v2.2.0
 	github.com/fsnotify/fsnotify v1.7.0
 	github.com/niklasfasching/go-org v1.7.0
@@ -16,7 +17,6 @@ require (
 )
 
 require (
-	github.com/alecthomas/kong v0.8.1 // indirect
 	github.com/osteele/tuesday v1.0.3 // indirect
 	github.com/tdewolff/parse/v2 v2.7.11 // indirect
 	golang.org/x/sys v0.16.0 // indirect
